code works however embedded-test pulls in std for some reason? Which then shoots the binary size for the tests. Will have to split them up and duplicate
